Inline decorator-kode in all form intropage templates. Otherwise, they won't be included in the example code.
