--- aksel.nav.no/website/components/utils/dekorator.tsx
@@ -1,46 +1,0 @@
-import React, { useEffect } from "react";
-
-function Header() {
-  return <div id="decorator-header" />;
-}
-
-function Footer() {
-  return <div id="decorator-footer" />;
-}
-
-const MILJO_URL = "https://www.nav.no/dekoratoren";
-
-function Env() {
-  return (
-    <div
-      id="decorator-env"
-      data-src={`${MILJO_URL}/env?context=privatperson&simple=true&availableLanguages=[{"locale":"nb","url":"https://www.nav.no/person/kontakt-oss"},{"locale":"en","url":"https://www.nav.no/person/kontakt-oss/en/"}]`}
-    />
-  );
-}
-
-/**
- * OBS: Dette er ikke anbefalt metode for å laste dekoratør!
- * Se `nav-dekoratoren`-dokumentasjon for riktig implementasjon
- * @see https://github.com/navikt/nav-dekoratoren
- */
-function useDekorator() {
-  useEffect(() => {
-    const script = document.createElement("script");
-    script.src = `${MILJO_URL}/client.js`;
-    script.async = true;
-    document.body.appendChild(script);
-
-    const styles = document.createElement("link");
-    styles.href = `${MILJO_URL}/css/client.css`;
-    styles.rel = "stylesheet";
-    document.head.appendChild(styles);
-
-    return () => {
-      document.body.removeChild(script);
-      document.head.removeChild(styles);
-    };
-  }, []);
-}
-
-export { Header, Footer, Env, useDekorator };
