+++ aksel.nav.no/website/pages/templates/soknad-introside/intro-med-alternativ-soknad.tsx
@@ -1,3 +1,4 @@
+import { useEffect } from "react";
 import { ArrowRightIcon } from "@navikt/aksel-icons";
 import {
   Accordion,
@@ -17,12 +18,6 @@ import {
   VStack,
 } from "@navikt/ds-react";
 import { ApplicationPictogram } from "@/assets/Icons";
-import {
-  Env,
-  Footer,
-  Header,
-  useDekorator,
-} from "../../../components/utils/dekorator";
 
 function Example() {
   useDekorator();
@@ -226,6 +221,49 @@ function Example() {
       <Env />
     </Page>
   );
+}
+
+function Header() {
+  return <div id="decorator-header" />;
+}
+
+function Footer() {
+  return <div id="decorator-footer" />;
+}
+
+const MILJO_URL = "https://www.nav.no/dekoratoren";
+
+function Env() {
+  return (
+    <div
+      id="decorator-env"
+      data-src={`${MILJO_URL}/env?context=privatperson&simple=true&availableLanguages=[{"locale":"nb","url":"https://www.nav.no/person/kontakt-oss"},{"locale":"en","url":"https://www.nav.no/person/kontakt-oss/en/"}]`}
+    />
+  );
+}
+
+/**
+ * OBS: Dette er ikke anbefalt metode for å laste dekoratør!
+ * Se `nav-dekoratoren`-dokumentasjon for riktig implementasjon
+ * @see https://github.com/navikt/nav-dekoratoren
+ */
+function useDekorator() {
+  useEffect(() => {
+    const script = document.createElement("script");
+    script.src = `${MILJO_URL}/client.js`;
+    script.async = true;
+    document.body.appendChild(script);
+
+    const styles = document.createElement("link");
+    styles.href = `${MILJO_URL}/css/client.css`;
+    styles.rel = "stylesheet";
+    document.head.appendChild(styles);
+
+    return () => {
+      document.body.removeChild(script);
+      document.head.removeChild(styles);
+    };
+  }, []);
 }
 
 // EXAMPLES DO NOT INCLUDE CONTENT BELOW THIS LINE
